+++ NEWS.md
@@ -21,6 +21,8 @@
 - .read_raster_source now uses a single logic to cover all cases (e.g. single tiles,
   tiled rasters with and without temporal dimension, single temporal rasters)
 - .read_raster_source now projects assets in case their CRS differs from the raster
+- .read_raster_source now applies a precision round-trip of 4 decimal places to
+  match rasters with slight changes in their spatial extent (#217, #218)
 
 
 # mapme.biodiversity 0.4.0